# docs/swagger/swagger.yaml
basePath: /
definitions:
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
...
  dto.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.AuthResponse:
    properties:
      access_token:
        type: string
    type: object
...
  dto.PublicReserveResponse:
    properties:
      status:
        type: string
    type: object
  dto.PublicWishlistItem:
    properties:
      description:
        type: string
      id:
        type: string
      is_reserved:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      product_link:
        type: string
      reserved_at:
        type: string
    type: object
  dto.PublicWishlistResponse:
    properties:
      description:
        type: string
      event_date:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.PublicWishlistItem'
        type: array
      name:
        type: string
    type: object
...
  response.ErrorPayload:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: 'REST API для вишлистов: авторизация, CRUD вишлистов/подарков, публичный
    просмотр и резервирование.'
  title: Wishlist Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Вход пользователя по email и паролю.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрация пользователя по email и паролю.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Register user
      tags:
      - Auth
  /public/{token}:
    get:
      description: Возвращает публичный вишлист с позициями по токену. Авторизация
        не требуется.
      parameters:
      - description: Public token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PublicWishlistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Get public wishlist
      tags:
      - Public
  /public/{token}/reserve/{itemID}:
    post:
      description: Резервирует подарок по публичному токену вишлиста и ID позиции.
        Авторизация не требуется.
      parameters:
      - description: Public token
        in: path
        name: token
        required: true
        type: string
      - description: Item ID (UUID)
        in: path
        name: itemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PublicReserveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      summary: Reserve public wishlist item
      tags:
      - Public
  /users/me:
    get:
      description: Возвращает текущего авторизованного пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token with Bearer prefix. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"